Use an internal aligned buffer for packets
This is the fastest unaligned access solution. It minimizes unaligned
access and reduces the number of inner loop conditionals to one. If a
target doesn't support unaligned access, performance won't suffer much
because memcpy() is called on big (packet length) chunks. The maximum
length of a packet is now limited to DVBCSA_BS_MAX_PACKET_LEN (184
bytes).